Start adding support for multiple displays.
"display" is the abstraction that lets me have a normal playback
view, and separate overlays for things like volume level, seek
offset, playlist management and such.  These things are rendered
separately, and one of these displays is selected for display at any
given time.  This way, the user can be changing the volume during
playback and see the volume level overlay for a moment, and the
script can keep updating the track position in the background.

And while I'm in there, fix an off-by-one bug in the display
renderer code that did not allow the right-most column of pixels to
ever be used.