Add a "change volume" dialog.
This dialog pops up when you change the volume, showing the volume
number and a nice bar across the bottom.  The dialog lasts for 2
seconds, and then deselects itself (going back to the playback
display).

Split out the bar-rendering stuff from update_track_pos() into a
standalone function, use it for both the playback bar and the volume
bar.  Use common setup code for both of the display structures.