Add some input checking to render_bar_at.
The "$value" argument should be between 0 and 1, inclusive.  If passed
a value outside this range, the call to GD->fill() could exceed the
bounds of the bar, and clobber the display.  (This can happen when the
header parser rounds down the track length field, or screws up the VBR
calculation.)

Enforce the range to prevent this from happening.